requirements.txt
certifi==2018.8.13
Django==2.0.5
django-cors-headers==2.4.0
django-tastypie==0.14.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
pip==18.0
psycopg2==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
setuptools==40.0.0
six==1.11.0
wheel==0.31.1
